Fixing transaction list bug, updating UI, and added logout feature

+++ app/src/main/java/com/evanedsaazola/ocbchomework/data/model/BalanceItem.kt
@@ -8,5 +8,5 @@ data class BalanceItem(
     @SerializedName("accountNo")
     val accountNo: String?,
     @SerializedName("balance")
-    val balance: Int?
+    val balance: Double?
 )
